+++ lib/core/design_system/components/task_card.dart
@@ -9,6 +9,7 @@ import '../theme/helper/app_assets.dart';
 import '../theme/helper/extention_colors.dart';
 import '../theme/helper/theme_extention.dart';
 import 'custom_btn_low_high_medium.dart';
+import 'horizontal_stacked_avatars.dart';
 
 class TaskState extends StatelessWidget{
   final String title;
@@ -147,7 +148,6 @@ class TaskCardItem extends StatelessWidget {
     );
   }
 }
-
 class ProgressBar extends StatelessWidget {
   final Color statusColor;
   final double taskStatus;
@@ -169,74 +169,6 @@ class ProgressBar extends StatelessWidget {
         borderRadius: BorderRadius.circular(4),
         color: statusColor,
         minHeight: 4,
-      ),
-    );
-  }
-}
-
-class HorizontalStackedAvatars extends StatelessWidget {
-  const HorizontalStackedAvatars({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    const double avatarRadius = 12.0;
-    const double overlap = 6.0;
-
-    return SizedBox(
-      width: (avatarRadius * 2 * 3) - (overlap * 2),
-      height: avatarRadius * 2,
-      child: Stack(
-        clipBehavior: Clip.none,
-        children: [
-          Positioned(
-            left: 0,
-            child: FunCircleAvatar(
-              image: const AssetImage(AppAssets.profilePlaceholder),
-            ),
-          ),
-          Positioned(
-            left: avatarRadius * 2 - overlap,
-            child: FunCircleAvatar(
-              image: const AssetImage(AppAssets.profilePlaceholder),
-            ),
-          ),
-          Positioned(
-            left: (avatarRadius * 2 - overlap) * 2,
-            child: FunCircleAvatar(
-              image: const AssetImage(AppAssets.profilePlaceholder),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class FunCircleAvatar extends StatelessWidget {
-  final ImageProvider image;
-  final double radius;
-  final double borderWidth;
-
-  const FunCircleAvatar({
-    super.key,
-    required this.image,
-    this.radius = 12.0,
-    this.borderWidth = 1.0,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        border: Border.all(
-          color: Colors.white,
-          width: borderWidth,
-        ),
-      ),
-      child: CircleAvatar(
-        radius: radius,
-        backgroundImage: image,
       ),
     );
   }
